+++ standard_logger.go
@@ -7,7 +7,7 @@ func NewStandardLogger(logger Logger, level Level, prefix string, flag int) *log
 	return log.New(NewLevelWriter(logger, level), prefix, flag)
 }
 
-// NewStandardErrorLogger returns a new standard library logger for error level logging (eg. for HTTP servers).
-func NewStandardErrorLogger(logger Logger, prefix string, flag int) *log.Logger {
+// NewErrorStandardLogger returns a new standard library logger for error level logging (eg. for HTTP servers).
+func NewErrorStandardLogger(logger Logger, prefix string, flag int) *log.Logger {
 	return NewStandardLogger(logger, Error, prefix, flag)
 }